financeflow/backend/requirements.txt

fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
alembic==1.13.3
